fix: resolve all ESLint violations and clean up PR #753
- Extract server helpers to fix max-lines violations:
  - streamUsageHandler.js (from streamEventHandler.js)
  - canvas-helpers.js (from canvas.js)
  - projects-session-helpers.js (from projects.js)
  - sessions-lifecycle.js (from sessions.js)
- Extract frontend modules to fix max-lines violations:
  - sessions/workLogActions.js (from sessions.js store)
  - utils/clipboard.js (from CommandButtonItem.vue)
- Compact CSS in Vue components to fix max-lines:
  - CanvasFileViewerHeader, CommandButtonItem, SummaryTab,
    NewSessionView, SessionDetailView
- Remove all debug/diagnostic logging ([DEBUG], [MODEL AUDIT])
- Remove plan-stricter-lint-and-coverage.md from repo root
- Fix max-nested-callbacks in sessions.test.js and todos.test.js
- Remove dead code (toggleStarredFilter, redundant cardIndex check)

All 7,488 tests pass. ESLint reports 0 errors, 0 warnings.
