+++ packages/server/src/services/streamEventHandler.js
@@ -4,8 +4,13 @@ import { WS_MESSAGE_TYPES } from '@claudetools/shared';
 import { updateTodos } from './todoStore.js';
 import * as summaryService from './summaryService.js';
 import * as diffService from './diffService.js';
-import { updateTurnUsage, currentTurnUsage, estimatedOutputTokens, estimateTokens } from './usageTracker.js';
 import * as kanbanService from './kanbanService.js';
+import {
+  handleMessageStart,
+  handleMessageDelta,
+  handleTextDelta as _handleTextDelta,
+  handleResultUsage,
+} from './streamUsageHandler.js';
 
 // ── Shared module-level state ──────────────────────────────────────────────
 
@@ -136,24 +141,15 @@ function handleSystemEvent(sessionId, event) {
   // Store Claude's session info
   if (event.subtype !== 'init') return;
 
-  // [MODEL AUDIT] Log model reported by SDK in system.init
-  console.log(`[MODEL AUDIT - SDK Event] system.init received:`, {
-    sessionId,
-    sdkSessionId: event.session_id,
-    modelFromSDK: event.model,
-  });
-
   // Save Claude session ID to the active conversation for context isolation
   const activeConversation = conversations.getActiveBySessionId(sessionId);
   if (activeConversation) {
     conversations.update(activeConversation.id, {
       claudeSessionId: event.session_id,
     });
-    console.log(`[MODEL AUDIT - SDK Event] Updated conversation ${activeConversation.id} claudeSessionId to ${event.session_id}`);
   }
   // Track current model for this session (used when creating messages)
   currentModels.set(sessionId, event.model);
-  console.log(`[MODEL AUDIT - SDK Event] Set currentModels[${sessionId}] = "${event.model}"`);
   // Capture available slash commands (do NOT update model here — session.model
   // tracks the user-requested short format; this SDK model is stored in currentModels)
   sessions.update(sessionId, {
@@ -209,11 +205,7 @@ function handleAssistantTextContent(sessionId, textContent, toolUseBlocks) {
   const activeConversation = conversations.getActiveBySessionId(sessionId);
   const conversationId = activeConversation?.id || null;
   const currentModel = currentModels.get(sessionId) || null;
-  // [MODEL AUDIT] Log model being saved with message
-  console.log(`[MODEL AUDIT - Message Save] Creating assistant message with model: "${currentModel}"`);
   const message = messages.create(sessionId, 'assistant', textContent, { toolUse, conversationId, model: currentModel });
-  console.log(`[MODEL AUDIT - Message Save] Created message ${message.id} in conversation ${conversationId} with model: "${currentModel}"`);
-  console.log(`[SESSION] assistant event: created assistant message ${message.id} in conversation ${conversationId} with model ${currentModel}`);
 
   // Associate pending work logs with this message immediately
   // This ensures work logs are attached to the correct message, not just the last one
@@ -227,7 +219,6 @@ function handleAssistantTextContent(sessionId, textContent, toolUseBlocks) {
     message,
     conversationId, // Include conversation context to prevent ambiguity
   });
-  console.log(`[SESSION] assistant event: broadcast assistant message ${message.id} to conversation ${conversationId}`);
 
   // Clear partial text on client now that complete message has been sent
   broadcastToSession(sessionId, WS_MESSAGE_TYPES.SESSION_PARTIAL, {
@@ -320,44 +311,14 @@ function handleStreamEventType(sessionId, event) {
 }
 
 /**
- * Handle stream_event > message_start — initial usage (input tokens)
+ * Handle stream_event > message_start — clear text accumulator, then delegate to streamUsageHandler
  * @param {string} sessionId
  * @param {Object} event
  */
-function handleMessageStart(sessionId, event) {
+function handleMessageStartWrapper(sessionId, event) {
   // Clear text accumulator for fresh message
   textAccumulators.delete(sessionId);
-
-  const usage = event.event?.message?.usage;
-  if (usage) {
-    const conversationId = activeConversationIds.get(sessionId);
-    const turnUsage = updateTurnUsage(conversationId, usage, 'message_start');
-    broadcastToSession(sessionId, WS_MESSAGE_TYPES.SESSION_USAGE_UPDATE, {
-      sessionId,
-      conversationId,
-      usage: turnUsage,
-      isFinal: false,
-    });
-  }
-}
-
-/**
- * Handle stream_event > message_delta — streaming output tokens
- * @param {string} sessionId
- * @param {Object} event
- */
-function handleMessageDelta(sessionId, event) {
-  const usage = event.event?.usage;
-  if (usage) {
-    const conversationId = activeConversationIds.get(sessionId);
-    const turnUsage = updateTurnUsage(conversationId, usage, 'message_delta');
-    broadcastToSession(sessionId, WS_MESSAGE_TYPES.SESSION_USAGE_UPDATE, {
-      sessionId,
-      conversationId,
-      usage: turnUsage,
-      isFinal: false,
-    });
-  }
+  handleMessageStart(sessionId, event);
 }
 
 /**
@@ -369,63 +330,13 @@ function handleContentBlockDelta(sessionId, event) {
   const delta = event.event.delta;
 
   if (delta?.type === 'text_delta' && delta.text) {
-    handleTextDelta(sessionId, delta);
+    _handleTextDelta(sessionId, delta, textAccumulators);
   }
 
   // Handle thinking delta - accumulate and broadcast partial (don't create work log yet)
   if (delta?.type === 'thinking_delta' && delta.thinking) {
     handleThinkingDelta(sessionId, delta);
   }
-}
-
-/**
- * Handle text_delta within content_block_delta — accumulate text and estimate tokens
- * @param {string} sessionId
- * @param {Object} delta
- */
-function handleTextDelta(sessionId, delta) {
-  // Accumulate text content
-  const current = textAccumulators.get(sessionId) || '';
-  const accumulated = current + delta.text;
-  textAccumulators.set(sessionId, accumulated);
-
-  broadcastToSession(sessionId, WS_MESSAGE_TYPES.SESSION_PARTIAL, {
-    sessionId,
-    text: accumulated,
-  });
-
-  // Estimate tokens from streamed content for real-time output token updates
-  const conversationId = activeConversationIds.get(sessionId);
-  if (!conversationId) return;
-
-  const currentEstimate = estimatedOutputTokens.get(conversationId) || 0;
-  const newEstimate = currentEstimate + estimateTokens(delta.text);
-  estimatedOutputTokens.set(conversationId, newEstimate);
-
-  // Get current turn usage and add estimated output
-  const turnData = currentTurnUsage.get(conversationId) || {
-    inputTokens: 0,
-    outputTokens: 0,
-    lastMessageOutput: 0,
-    cacheReadInputTokens: 0,
-    cacheCreationInputTokens: 0,
-  };
-
-  // Broadcast usage update with estimated tokens
-  const broadcastUsage = {
-    inputTokens: turnData.inputTokens,
-    outputTokens: turnData.outputTokens + Math.max(turnData.lastMessageOutput, newEstimate),
-    cacheReadInputTokens: turnData.cacheReadInputTokens,
-    cacheCreationInputTokens: turnData.cacheCreationInputTokens,
-  };
-
-  broadcastToSession(sessionId, WS_MESSAGE_TYPES.SESSION_USAGE_UPDATE, {
-    sessionId,
-    conversationId,
-    usage: broadcastUsage,
-    isFinal: false,
-    isEstimate: true,  // Flag so UI can show "~" prefix if desired
-  });
 }
 
 /**
@@ -518,157 +429,12 @@ function handleResultSuccess(sessionId, event) {
   }
 }
 
-/**
- * Handle final usage stats from result event — update conversation and session usage
- * @param {string} sessionId
- * @param {Object} event
- */
-function handleResultUsage(sessionId, event) {
-  const turnUsage = extractTurnUsage(sessionId, event);
-
-  // [MODEL AUDIT] Log model from result event
-  console.log(`[MODEL AUDIT - Result Event] Turn usage model extraction:`, {
-    modelUsageKeys: Object.keys(event.modelUsage || {}),
-    primaryModelFromInit: currentModels.get(sessionId),
-    extractedModel: turnUsage.model,
-    rawModelUsage: event.modelUsage,
-  });
-
-  // Get the conversation ID for this session's current turn
-  const conversationId = activeConversationIds.get(sessionId);
-  const currentConversation = conversationId ? conversations.getById(conversationId) : null;
-
-  // Update conversation with cumulative usage (add to existing)
-  const updatedConversation = updateConversationUsage(conversationId, currentConversation, turnUsage);
-
-  // Also update session-level usage (aggregate of all conversations) for backward compatibility
-  const cumulativeSessionUsage = buildCumulativeSessionUsage(sessionId, turnUsage);
-  const updatedSession = sessions.updateUsage(sessionId, cumulativeSessionUsage);
-
-  // Broadcast final usage update with conversationId
-  broadcastToSession(sessionId, WS_MESSAGE_TYPES.SESSION_USAGE_UPDATE, {
-    sessionId,
-    conversationId,
-    usage: updatedConversation ? {
-      inputTokens: updatedConversation.inputTokens,
-      outputTokens: updatedConversation.outputTokens,
-      cacheReadInputTokens: updatedConversation.cacheReadInputTokens,
-      cacheCreationInputTokens: updatedConversation.cacheCreationInputTokens,
-      webSearchRequests: updatedConversation.webSearchRequests,
-      contextWindow: updatedConversation.contextWindow,
-    } : cumulativeSessionUsage,
-    turnUsage,
-    isFinal: true,
-  });
-
-  // Also broadcast session update for session list
-  broadcastToProject(updatedSession.projectId, WS_MESSAGE_TYPES.SESSION_UPDATED, {
-    projectId: updatedSession.projectId,
-    sessionId,
-    session: updatedSession,
-  });
-
-  // Broadcast conversation update for real-time UI updates
-  if (updatedConversation) {
-    broadcastToSession(sessionId, WS_MESSAGE_TYPES.CONVERSATION_UPDATED, {
-      sessionId,
-      conversation: updatedConversation,
-    });
-  }
-
-  // Clean up turn usage and estimated tokens
-  currentTurnUsage.delete(conversationId);
-  estimatedOutputTokens.delete(conversationId);
-  activeConversationIds.delete(sessionId);
-}
-
-/**
- * Extract turn usage from result event's modelUsage or usage fields
- * @param {string} sessionId
- * @param {Object} event
- * @returns {Object} turnUsage
- */
-function extractTurnUsage(sessionId, event) {
-  // Extract from modelUsage if available (has more detail)
-  const modelUsageEntry = event.modelUsage
-    ? Object.values(event.modelUsage)[0]
-    : null;
-
-  // Use the model from system.init (stored in currentModels) rather than modelUsage keys
-  // because modelUsage can contain multiple models when sub-agents are used (e.g., Opus using Haiku)
-  // and Object.keys()[0] would pick the wrong model
-  const primaryModel = currentModels.get(sessionId) || Object.keys(event.modelUsage || {})[0] || null;
-
-  return {
-    inputTokens: modelUsageEntry?.inputTokens || event.usage?.input_tokens || 0,
-    outputTokens: modelUsageEntry?.outputTokens || event.usage?.output_tokens || 0,
-    cacheReadInputTokens: modelUsageEntry?.cacheReadInputTokens || event.usage?.cache_read_input_tokens || 0,
-    cacheCreationInputTokens: modelUsageEntry?.cacheCreationInputTokens || event.usage?.cache_creation_input_tokens || 0,
-    webSearchRequests: modelUsageEntry?.webSearchRequests || 0,
-    contextWindow: modelUsageEntry?.contextWindow || 200000,
-    model: primaryModel,
-  };
-}
-
-/**
- * Build cumulative session-level usage by adding turn usage to existing session usage
- * @param {string} sessionId
- * @param {Object} turnUsage
- * @returns {Object} cumulativeSessionUsage
- */
-function buildCumulativeSessionUsage(sessionId, turnUsage) {
-  const currentSession = sessions.getById(sessionId);
-  return {
-    inputTokens: (currentSession.inputTokens || 0) + turnUsage.inputTokens,
-    outputTokens: (currentSession.outputTokens || 0) + turnUsage.outputTokens,
-    cacheReadInputTokens: (currentSession.cacheReadInputTokens || 0) + turnUsage.cacheReadInputTokens,
-    cacheCreationInputTokens: (currentSession.cacheCreationInputTokens || 0) + turnUsage.cacheCreationInputTokens,
-    webSearchRequests: (currentSession.webSearchRequests || 0) + turnUsage.webSearchRequests,
-    contextWindow: turnUsage.contextWindow,
-  };
-}
-
-/**
- * Update conversation with cumulative usage from turn
- * @param {string|undefined} conversationId
- * @param {Object|null} currentConversation
- * @param {Object} turnUsage
- * @returns {Object|null} updatedConversation
- */
-function updateConversationUsage(conversationId, currentConversation, turnUsage) {
-  if (!currentConversation) return null;
-
-  // [MODEL AUDIT] Log conversation model before update
-  console.log(`[MODEL AUDIT - Conversation Update] Before updateUsage:`, {
-    conversationId,
-    currentConversationModel: currentConversation.model,
-    newModelFromUsage: turnUsage.model,
-  });
-
-  const cumulativeConversationUsage = {
-    inputTokens: (currentConversation.inputTokens || 0) + turnUsage.inputTokens,
-    outputTokens: (currentConversation.outputTokens || 0) + turnUsage.outputTokens,
-    cacheReadInputTokens: (currentConversation.cacheReadInputTokens || 0) + turnUsage.cacheReadInputTokens,
-    cacheCreationInputTokens: (currentConversation.cacheCreationInputTokens || 0) + turnUsage.cacheCreationInputTokens,
-    webSearchRequests: (currentConversation.webSearchRequests || 0) + turnUsage.webSearchRequests,
-    contextWindow: turnUsage.contextWindow,
-  };
-
-  const updatedConversation = conversations.updateUsage(conversationId, cumulativeConversationUsage);
-  // [MODEL AUDIT] Log conversation model after update
-  console.log(`[MODEL AUDIT - Conversation Update] After updateUsage:`, {
-    conversationId,
-    updatedConversationModel: updatedConversation?.model,
-  });
-
-  return updatedConversation;
-}
 
 // ── Dispatch maps ───────────────────────────────────────────────────────────
 
 /** @type {Record<string, (sessionId: string, event: Object) => void>} */
 const streamSubHandlers = {
-  message_start: handleMessageStart,
+  message_start: handleMessageStartWrapper,
   message_delta: handleMessageDelta,
   content_block_delta: handleContentBlockDelta,
   content_block_stop: handleContentBlockStop,
