+++ packages/server/src/ws/WebSocketManager.js
@@ -111,7 +111,6 @@ export class WebSocketManager {
       this.#projectSubscriptions.set(projectId, new Set());
     }
     this.#projectSubscriptions.get(projectId).add(ws);
-    console.log(`🔷 [WS Manager] Client subscribed to project ${projectId}, total subscribers: ${this.#projectSubscriptions.get(projectId).size}`);
   }
 
   #handleUnsubscribeProject(ws, message) {
@@ -179,11 +178,6 @@ export class WebSocketManager {
    */
   broadcastToProject(projectId, type, payload) {
     const subscribers = this.#projectSubscriptions.get(projectId);
-    // ========== DIAGNOSTIC LOGGING ==========
-    if (type === 'command:run:complete' || type === 'command:run:output') {
-      console.log(`🔷 [WS Manager] broadcastToProject ${type} to ${projectId}, subscribers: ${subscribers?.size || 0}`);
-    }
-    // ========================================
     if (!subscribers || subscribers.size === 0) return;
 
     const message = createMessage(type, payload);
