+++ packages/web/src/utils/clipboard.js
@@ -1,0 +1,32 @@
+/**
+ * Copy text to clipboard using modern Clipboard API with fallback.
+ * @param {string} text - Text to copy
+ * @returns {Promise<boolean>} Whether the copy succeeded
+ */
+export async function copyToClipboard(text) {
+  // Try modern Clipboard API first
+  if (navigator.clipboard?.writeText) {
+    try {
+      await navigator.clipboard.writeText(text);
+      return true;
+    } catch (err) {
+      console.error('Clipboard API failed:', err);
+    }
+  }
+
+  // Fallback for older browsers
+  try {
+    const textarea = document.createElement('textarea');
+    textarea.value = text;
+    textarea.style.position = 'fixed';
+    textarea.style.opacity = '0';
+    document.body.appendChild(textarea);
+    textarea.select();
+    document.execCommand('copy');
+    document.body.removeChild(textarea);
+    return true;
+  } catch (fallbackErr) {
+    console.error('Fallback copy failed:', fallbackErr);
+    return false;
+  }
+}
